requirements.txt
asn1crypto==0.24.0
async-generator==1.10
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
future==0.17.1
idna==2.7
pyaes==1.6.1
pyasn1==0.4.4
pycparser==2.19
PySocks==1.6.8
python-decouple==3.1
python-telegram-bot==11.1.0
requests==2.20.1
rsa==4.0
schedule==0.5.0
six==1.11.0
SQLAlchemy==1.2.14
Telethon-sync==1.1.1
urllib3==1.24.1
